requirements.txt
fastapi
uvicorn
motor
passlib
PyJWT
fastapi-pagination
python-dotenv
pydantic
aiofiles
python-multipart
werkzeug